break index.tsx into multiple files
