+++ ios/Hash.swift
