+++ src/index.ts
@@ -1,0 +1,18 @@
+import { NativeModules, Platform } from 'react-native';
+
+const LINKING_ERROR =
+  `The package 'react-native-wallet-core' doesn't seem to be linked. Make sure: \n\n` +
+  Platform.select({ ios: "- You have run 'pod install'\n", default: '' }) +
+  '- You rebuilt the app after installing the package\n' +
+  '- You are not using Expo managed workflow\n';
+
+const ReactNativeWalletCore = NativeModules.ReactNativeWalletCore ? NativeModules.ReactNativeWalletCore : new Proxy(
+  {},
+  {
+    get() {
+      throw new Error(LINKING_ERROR);
+    },
+  }
+);
+
+export { ReactNativeWalletCore };
